Moved Filter classes into their own package to not pollute file system. Added FilterDepartment to hand inputs such as COMP or MECE to filter by
