+++ src/main/java/edu/gcc/comp350/jmeg/filter/FilterYear.java
@@ -1,4 +1,6 @@
-package edu.gcc.comp350.jmeg;
+package edu.gcc.comp350.jmeg.filter;
+
+import edu.gcc.comp350.jmeg.Course;
 
 import java.util.ArrayList;
 import java.util.stream.Collectors;
